Revert "Aligned NonbondedInteractionGroup and NonbondedAllPairs potentials (#1538)"
This reverts commit 6c3a31e834163800148f1bed476c4afe47b195a6.
