+++ tests/test_single_topology_combined.py
@@ -216,6 +216,4 @@ def test_nonbonded_host_params_independent_of_lambda(
     params0 = get_nonbonded_host_params(0.0)
     for lamb in np.linspace(0.1, 1, 10):
         params = get_nonbonded_host_params(lamb)
-        np.testing.assert_array_equal(
-            params.astype(np.float32).astype(np.float64), params0.astype(np.float32).astype(np.float64)
-        )
+        np.testing.assert_array_equal(params, params0)
